pyproject.toml
[project]
name = "sam3-annotation-tool"
version = "0.1.0"
description = "Gradio interface for SAM3 image annotation"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "transformers @ git+https://github.com/huggingface/transformers.git@7be2d3587efede549bfde514ffb7a00e6a9baa20",
    "sentencepiece==0.2.1",
    "opencv-python==4.12.0.88",
    "imageio[pyav]==2.37.2",
    "torchvision==0.24.1",
    "matplotlib==3.10.8",
    "accelerate==1.12.0",
    "kernels==0.11.7",
    "pillow==12.1.0",
    "gradio==6.0.2",
    "spaces==0.46.0",
    "numpy==2.2.6",
    "torch==2.9.1",
    "peft==0.18.1",
    "shapely==2.1.2"
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.metadata]
allow-direct-references = true

[tool.hatch.build.targets.wheel]
packages = ["src"]

[dependency-groups]
dev = [
    "pytest>=9.0.2",
]
